+++ reporter/apps.py
@@ -7,8 +7,9 @@ from typing import Optional
 from slack import WebClient
 
 from .adapters import JiraAdapter
-from .conf import SLACK_CHANNEL_ID, SLACK_TOKEN, JIRA_SPRINT
+from .conf import JIRA_SPRINT, SLACK_CHANNEL_ID, SLACK_TOKEN
 from .slughify import slughifi
+from .utils import get_value_from_redis, set_key_in_redis
 
 __version__ = '0.0.4'
 
@@ -40,16 +41,12 @@ class SlackApp:
     def __init__(self, **kwargs):
         """Initialize."""
         loop = kwargs.get('loop') or asyncio.get_event_loop()
+
         self.version = f'*version:* {__version__}'
         self.channel_id = kwargs.get('channel_id') or SLACK_CHANNEL_ID
-        self.slack = WebClient(token=SLACK_TOKEN, loop=loop, run_async=True)
-        self.known_user_ids = {}
+        self.known_user_ids = get_value_from_redis('slack-known-user-ids') or {}
 
-    async def send_no_pull_requests_message(self) -> None:
-        """Send a default message when no pull requests."""
-        message = self._render_template('no_pull_requests.json')
-        message['blocks'][1]['elements'][1]['text'] = self.version
-        await self.send_message(message)
+        self.slack = WebClient(token=SLACK_TOKEN, loop=loop, run_async=True)
 
     @staticmethod
     def _render_template(filename: str) -> dict:
@@ -65,9 +62,6 @@ class SlackApp:
 
         :param pull_requests: information about the pull requests
         """
-        users = await self.slack.users_list()
-        users = users['members']
-
         header = self._render_template('header.json')
         author = self._render_template('autor.json')
         author['elements'][1]['text'] = self.version
@@ -87,7 +81,7 @@ class SlackApp:
             for pull_request in issue['pull_requests']:
                 description = deepcopy(description_block)
                 reviewers = map(
-                    lambda name: self._get_user_mention(name, users),
+                    lambda name: self._get_user_mention(name),
                     pull_request['reviewers'],
                 )
                 description['text']['text'] = ' '.join(reviewers)
@@ -98,9 +92,10 @@ class SlackApp:
                 title['text']['text'] = f':bender: *[{issue["key"]}] {issue["title"]}*'
                 message['blocks'].extend([title] + descriptions + [divider])
 
+        set_key_in_redis('slack-known-user-ids', self.known_user_ids)
         await self.send_message(message)
 
-    def _get_user_mention(self, name: str, users: list) -> str:
+    def _get_user_mention(self, name: str) -> str:
         """
         Get text for a user mention.
 
@@ -108,7 +103,6 @@ class SlackApp:
         it will return his name.
 
         :param name: a user's name from a pull request
-        :param users: users list from slack
 
         :returns: a mention string
         """
@@ -118,7 +112,7 @@ class SlackApp:
             pass
 
         mention = name
-        user = self._get_user(name, users)
+        user = self._get_user(name, get_value_from_redis('slack-members'))
         if user:
             mention = f'<@{user["id"]}>'
 
@@ -142,6 +136,12 @@ class SlackApp:
                 return self._get_user(name, users, first_time=False)
             return None
         return user
+
+    async def send_no_pull_requests_message(self) -> None:
+        """Send a default message when no pull requests."""
+        message = self._render_template('no_pull_requests.json')
+        message['blocks'][1]['elements'][1]['text'] = self.version
+        await self.send_message(message)
 
     async def send_message(self, message: dict) -> None:
         """