+++ server/tests/test_handlers.py
@@ -31,7 +31,7 @@ class HomeHandlerTestCase(AsyncHTTPTestCase):
         response = self.fetch('/')
 
         self.assertEqual(response.code, 200)
-        self.assertEqual(response.body.decode('utf-8'), 'Hello world!')
+        self.assertEqual(response.body.decode(), 'Hello world!')
 
 
 class SlackHandlerTestCase(AsyncHTTPTestCase):
@@ -61,7 +61,7 @@ class SlackHandlerTestCase(AsyncHTTPTestCase):
     def _prepare_headers(self, body: dict) -> dict:
         """Prepare request headers."""
         timestamp = str(int(time()))
-        body = json.dumps(body).encode('utf-8')
+        body = json.dumps(body).encode()
         req = str.encode('v0:' + timestamp + ':') + body
         request_hash = 'v0=' + hmac.new(
             str.encode(self.signing_secret),
@@ -88,7 +88,7 @@ class SlackHandlerTestCase(AsyncHTTPTestCase):
         )
 
         self.assertEqual(response.code, 200)
-        self.assertEqual(response.body.decode('utf-8'), data['challenge'])
+        self.assertEqual(response.body.decode(), data['challenge'])
 
     def test_post_returns_forbidden_when_request_timestamp_is_not_given(self):
         """Test post returns 403 when there is no X-Slack-Request-Timestamp header."""