+++ CHANGELOG.rst
@@ -8,10 +8,12 @@ CHANGELOG
 **Updates**
 
 1. Moved filtering pull requests to parsers. From now on the parser returns only
-pull requests that have reviewers that did not give an approval.
+   pull requests that have reviewers that did not give an approval.
 
 2. Changed the pull requests message title and ticket title.
 
+3. Caching workspace users in Redis for performance. Also added a periodic task
+   to update that list every 24hours.
 
 0.0.3 (23.03.2020)
 ------------------